Move files and fix types
