+++ package.json
@@ -54,6 +54,8 @@
     "@testing-library/react": "^10.4.5",
     "@testing-library/user-event": "^12.0.11",
     "@types/enzyme": "^3.10.5",
+    "@types/jest": "^26.0.4",
+    "@types/mocha": "^8.0.0",
     "@types/react": "^16.9.43",
     "@types/react-dom": "^16.9.8",
     "babel-core": "^7.0.0-bridge.0",
@@ -108,7 +110,8 @@
   ],
   "jest": {
     "roots": [
-      "<rootDir>/src"
+      "<rootDir>/src",
+      "<rootDir>/test"
     ],
     "collectCoverageFrom": [
       "src/**/*.{js,jsx,ts,tsx}",
@@ -124,8 +127,8 @@
       "<rootDir>/config/setupTests.js"
     ],
     "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-      "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
+      "<rootDir>/**/__tests__/**/*.{js,jsx,ts,tsx}",
+      "<rootDir>/**/*.{spec,test}.{js,jsx,ts,tsx}"
     ],
     "testEnvironment": "jest-environment-jsdom-fourteen",
     "testURL": "http://localhost",
