+++ src/PlyComponent.examples.md
